// apps/api/prisma/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const LearningStatus = {
  SAVED: 'SAVED',
  NEW: 'NEW',
  LEARNING: 'LEARNING',
  INCOMPLETE: 'INCOMPLETE',
  PROFICIENT: 'PROFICIENT',
  MASTERED: 'MASTERED'
} as const

export type LearningStatus = (typeof LearningStatus)[keyof typeof LearningStatus]


export const ExerciseType = {
  FLASHCARD: 'FLASHCARD',
  MULTIPLE_CHOICE: 'MULTIPLE_CHOICE',
  TYPING: 'TYPING',
  CLOZE: 'CLOZE',
  ARTICLE_GUESS: 'ARTICLE_GUESS'
} as const

export type ExerciseType = (typeof ExerciseType)[keyof typeof ExerciseType]